Fixing a bunch of warnings, deactivating offset-of waring because it's useless.

+++ src/gfx/Mesh.cpp
@@ -36,12 +36,12 @@ void Mesh::UploadToGfxDevice() {
 }
 
 void Mesh::Destroy() {
-	if (posVbo != -1) {
+	if ((int)posVbo != -1) {
 		glDeleteBuffers(1, &posVbo);
 		posVbo = -1;
 	}
 
-	if (uvVbo != -1) {
+	if ((int)uvVbo != -1) {
 		glDeleteBuffers(1, &uvVbo);
 		uvVbo = -1;
 	}
